Added: get current users, suspend them, revert their status from suspended to normal & Refactored: admin_signin to admin_api, moved list_new_users view from register to admin_api
