+++ mbtb_app/resources/apis/users/register/urls.py
@@ -3,7 +3,6 @@ from . import views
 from rest_framework import routers
 
 router = routers.DefaultRouter()
-router.register('list_new_users', views.NewUsersListViewSet)
 router.register('add_new_users', views.NewUsersViewSet)
 
 urlpatterns = [
