+++ mbtb_app/resources/apis/users/register/views.py
@@ -1,14 +1,7 @@
 from rest_framework import viewsets
 from .models import Users
 from .serializers import UsersSerializer
-from .permissions import IsAuthenticated, IsPostAllowed
-
-
-# For authenticating admin credentials and return auth token
-class NewUsersListViewSet(viewsets.ModelViewSet):
-    permission_classes = [IsAuthenticated]
-    queryset = Users.objects.filter(pending_approval='Y')  # filtering to fetch only pending requests
-    serializer_class = UsersSerializer
+from permissions.is_post_allowed import IsPostAllowed
 
 
 # For new user registration requests and only post request allowed
