+++ mbtb_app/resources/apis/users/users/urls.py
@@ -19,6 +19,6 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('register.urls')),
-    path('', include('admin_signin.urls')),
+    path('', include('admin_api.urls')),
     path('', include('users_signin.urls')),
 ]
